Move utils functions used in metadata  to domain
