+++ src/domain/instance/Instance.ts
@@ -1,3 +1,4 @@
+import _ from "lodash";
 
 export interface InstanceData {
     id: string;
@@ -10,6 +11,7 @@ export interface InstanceData {
 }
 
 export default class Instance {
+
     private data: InstanceData;
 
     constructor(data: InstanceData) {
@@ -46,5 +48,9 @@ export default class Instance {
 
     public get apiVersion(): number | undefined {
         return this.version ? +this.version?.split(".")[1] : undefined
+    }
+
+    public toObject(): Omit<InstanceData, "password"> {
+        return _.omit(this.data, ["password"]);
     }
 }
