+++ src/domain/metadata/MetadataRepositoriy.ts
@@ -1,10 +1,10 @@
-import { MetadataImportResponse, MetadataImportParams } from "./Types";
+import { MetadataImportResponse, MetadataImportParams } from "./types";
 import Instance from "../instance/Instance";
 import { Id } from "../common/Entities";
 import { MetadataPackage, MetadataFieldsPackage, MetadataPackageSchema, MetadataEntity } from "./entities";
 
 export interface MetadataRepository {
-    getMetadataFieldsByIds<T>(ids: Id[], fields: string): Promise<MetadataFieldsPackage<T>>;
+    getMetadataFieldsByIds<T>(ids: Id[], fields: string, targetInstance?: Instance): Promise<MetadataFieldsPackage<T>>;
 
     getMetadataByIds(ids: Id[]): Promise<MetadataPackage>;
